requirements.txt
google-generativeai>=0.5.0
pillow
python-dotenv
flask
flask-socketio
gunicorn
eventlet
PyMuPDF
requests